do_function/monitor_snapshots/packages/forest-snapshots/main/requirements.txt
slackclient==2.9.4